novas modificaçãoes!

+++ src/controller/ServidorConvenioController.java
@@ -1,15 +1,21 @@
 package controller;
 
 import java.net.URL;
+import java.text.DecimalFormat;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import dao.ServidorDB;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -28,6 +34,8 @@ public class ServidorConvenioController implements Initializable {
 	private TableColumn<ServidorConvenio, Double> valorColuna;
 	@FXML
 	private Label lbTotal;
+	@FXML
+	private Button btnFechar;
 
 	@Override
 	public void initialize(URL arg0, ResourceBundle arg1) {
@@ -39,20 +47,28 @@ public class ServidorConvenioController implements Initializable {
 	@FXML
 	private void onActionRemover() {
 		ServidorConvenio c = tabela.getSelectionModel().getSelectedItem();
-		new ServidorDB().deleteConvenio(c.getNome());
-		update();
+		if (c != null) {
+			Alert a = new Alert(AlertType.CONFIRMATION);
+			a.setHeaderText("Remover Convenio do Servidor?");
+			a.getButtonTypes();
+			Optional<ButtonType> result = a.showAndWait();
 
+			if (result.get() == ButtonType.OK) {
+				new ServidorDB().deleteConvenio(c.getNome());
+				update();
+			}
+		}
 	}
 
 	@FXML
 	private void onActionAdicionar() {
-		TelaCadastroSC tccs = new TelaCadastroSC();
-		tccs.setServidor(servidor);
+		update();
+		TelaCadastroSC tcsc = new TelaCadastroSC();
+		tcsc.setServidor(servidor);
 		Stage stage = new Stage();
 		stage.centerOnScreen();
 		stage.initModality(Modality.APPLICATION_MODAL);
-		tccs.start(stage);
-		update();
+		tcsc.start(stage);
 	}
 
 	public void setServidor(Servidor servidor) {
@@ -60,6 +76,7 @@ public class ServidorConvenioController implements Initializable {
 		update();
 	}
 
+	@FXML
 	private void update() {
 		tabela.getItems().clear();
 		double valorTotal = 0.0;
@@ -69,8 +86,15 @@ public class ServidorConvenioController implements Initializable {
 			for (ServidorConvenio sc : lista) {
 				valorTotal += sc.getValor();
 			}
-			lbTotal.setText("Total: R$ " + valorTotal);
+			DecimalFormat df = new DecimalFormat("0.00");
+			lbTotal.setText("Total: R$ " + df.format(valorTotal));
 		}
 		tabela.refresh();
+	}
+
+	@FXML
+	private void fecharTela() {
+		Stage stage = (Stage) btnFechar.getScene().getWindow();
+		stage.close();
 	}
 }
